Gracefully handle child resources mutated by admission controllers
The API server is able to mutate resources as they are created and
updated. This is commonly used to apply defaults. When a controller
later compares the desired and actual state of the resource to determine
if it needs to be updated, these default values create the appearance of
a meaningful change that needs to be applied to the server.

When the update request is handled by the API server, the default value
is re-applied, creating effectively a no-op request. This is inefficient
because it creates unnecessary load on the API server.

Now, we compare the object as returned from an update request, capturing
the difference and applying that difference to the next update call. If
the object after having the diff applied is semantically equal to
current state, the update can be skipped.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>