Limit child patches to resources of the same generation
If the upstream resource is updated in a way that removes a field
contained in a patch, the patch should be invalidated and recreated. The
Generation is used as a proxy for a meaningful change in the target
resource. If the generations do not match, the diff is skipped and the
update attempt occurs as normal.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>