Validate sub reconcilers on setup
Because we use reflection to define callback methods with specific
types, we cannot relay on the go compiler to catch type errors.
Previously, an incompatible type would result in a panic at runtime. If
the method is lightly used, it may take time for it to be invoked and
crash the controller.

Now, the SetupWithManager method will return an appropriate error if the
reconciler is misconfigured (missing required value, wrong types, etc).

Breaking changes:
- ChildReconciler#ParentType is removed, the value is now passed from
  the the ParentReconciler. ChildReconciler and SyncReconcilers may use
  a struct or compatible interface to represent the parent resource.
- ChildReconciler#DesiredChild must accept a context as it's first
  argument. The context is used to pass state between sub reconcilers,
  always including will make it easier to retrieve stashed state.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>