Status only child reconciliation
There are times when there is insufficient state to create/update a
child resource, but an existing resource should not be removed. The
existing child's status should still be reflected on the parent.

To skip reconciliation of the child resource while still reflecting an
existing child's status on the parent, return OnlyReconcileChildStatus
as an error from the DesiredChild method.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>