Cast parent resource type
A CastParent reconciler casts the type of the ParentReconciler's type by
projecting the data onto a new resource. Casting the parent resource is
useful to create cross cutting reconcilers that can operate on common
portion of multiple parent resources, common referred to as a duck type.

JSON encoding is used as the intermediate representation. Operations on
a cast parent are read-only. Attempts to mutate the parent will result
in the reconciler erring.

```
func FunctionReconciler(c reconcilers.Config) *reconcilers.ParentReconciler {
	c.Log = c.Log.WithName("Function")

	return &reconcilers.ParentReconciler{
		Type: &buildv1alpha1.Function{},
		Reconciler: reconcilers.Sequence{
			&reconcilers.CastParent{
				Type: &duckv1alpha1.ImageRef{},
				Reconciler: &reconcilers.SyncReconciler{
					Sync: func(ctx context.Context, parent *duckv1alpha1.ImageRef) error {
						// do something with the duckv1alpha1.ImageRef instead of a buildv1alpha1.Function
						return nil
					},
					Config: c,
				},
			},
			FunctionChildImageReconciler(c),
		},

		Config: c,
	}
}
```

Signed-off-by: Scott Andrews <andrewssc@vmware.com>