Reflect mutations from cast parent back to parent (#59)
The CastParent higher order reconciler is useful for operating on a
fragment of a resource with a compatible structure represented by a duck
type. Previously, this mapping was read-only, with mutations of the cast
parent resulting in an error. Now changes to the cast parent are
reflected back to the original parent type.

Just like the cast resource will only contain fields defined by both the
parent type and the cast type, mutation on portions of the cast resource
that are not available on the parent type are dropped.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>