Drop ChildReconciler#IndexField
The IndexField was an internal implementation detail that leaked out.
It has been removed. Instead of using a field indexer we can manually
filter the list of child resources.

Besides leaking an implementation detail to users, field indexers must
be defined before an informer is started, blocking the lazy
initialization of a controller.

Breaking change:
- remove ChildReconciler IndexField, it is no longer needed

Signed-off-by: Scott Andrews <andrewssc@vmware.com>