Multiple ChildReconcilers for the same type within a single ParentReconciler
Multiple children of the same kind may not be reconciled by the same
parent reconciler. Each child need a separate instance of a
ChildReconciler. A new method OurChild, is used to distinguish between
children that should be controlled by the specific ChildReconciler.

The OurChild method should only be implemented when there are multiple
ChildReconcilers for the same type under the same ParentReconciler. A
child resource that doesn't match the OurChild method will be orphaned.
By default all children of the type owned by the parent are managed by
the ChildReconciler.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>