Track mutation patches when creating a resource
When a mutation webhook intercepts a request to create a resource, we
now capture the diff applied by the webhook. When a future turn through
the reconciler occurs, the patch is applied to the desired state before
an update occurs. Previously, the patch was only created from updates.

This change also refines how the generation is compared when creating
and applying the patch.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>