Replace custom conditions with metav1.Condition
This is a breaking change, although the transition is mostly a drop in.
The metav1.Condition type requires that a reason be provided for all
conditions. The severity field does not exist. The LastTransitionTime
value is metav1.Time instead of the custom VolitileTime.

The condition manager, reconciler test harnesses and factories have been
updated for the new condition type.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>