Add parent resource as a given
In a real controller the parent reconciler loads the parent resource
from the api server. When testing a sub reconciler, we can assume the
parent resource as a given resource.

Users who want to test when the parent resource is deleted from the api
server after the reconciliation starts can use a reactor to inject a
not found error.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>