Move config into context (#209)
Reconcilers and sub reconcilers use the config object to centralize
common api operations. Previously this config was passed to each
reconciler when it was initialized. However, this makes it harder to
dynamically swap the config at runtime based on the content of a
resource. For example, a resource may specify a service account to use
when creating other resources with least privilege.

The ChildReconciler and SyncReconciler no long allow the Config to be
specified directly. It must be passed in via the context. So that the
RetrieveConfig method can return the Config.

SubReconcilers that need to interact with the parent resource should use
the parent config via RetrieveParentConfig method. In many cases the
config and parent config are one in the same, however, the WithConfig
sub reconciler can inject an alternate config for the nested sub
reconcilers.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>