Add RBAC recommendations for controllers
The parent and child reconciler each make client requests based on the
resources being reconciled. We should give guidance to users as to what
RBAC permissions are required to effectively use the reconcilers.

The RBAC permissions are defined as kubebuilder annotations that can be
parsed and converted into a ClusterRole via controller-gen.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>