Handle alternate forms of resource status (#233)
We previously assumed that a resource either conformed to our notion of
status or didn't have a status field at all. If a status field existed,
but either didn't have the structure we expected or did have the
structure but was a pointer, would err.

Now we gracefully unpack the status as best we can, and safely ignore
cases where the structure does not conform to our ideal structure. In
those cases the functionality that would be applied is skipped. This
includes:
- calling `status.InitializeConditions()`
- setting `status.ObservedGeneration`
- normalizing the LastTransitionTime for conditions

When we can't  apply these behaviors, a warning is logged during
setup of the ParentReconciler.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>