AggregateReconciler
An AggregateReconciler is responsible for synthesizing a single
resource, aggregated from other state. The AggregateReconciler is a
fusion of the ResourceReconciler and ChildReconciler.

Instead of operating on all resources of a type, it will only operate on
a specific resource identified by the type and request (namespace and
name). Unlike the child reconciler, the "parent" and "child" resources
are the same.

The resource reconciler is responsible for:
- fetching the resource being reconciled
- creating a stash to pass state between sub reconcilers
- passing the resource to each sub reconciler in turn
- creates the resource if it does not exist
- updates the resource if it drifts from the desired state
- deletes the resource if no longer desired
- logging the reconcilers activities
- records events for mutations and errors

The implementor is responsible for:
- specifying the type, namespace and name of the aggregate resource
- defining the desired state
- indicating if two resources are semantically equal
- merging the actual resource with the desired state (often as simple as
  copying the spec and labels)

Signed-off-by: Scott Andrews <andrewssc@vmware.com>