Restructure ReconcilerTestSuite to a map
(Sub)ReconcilerTests is a map based equivalent to the slice based
(Sub)ReconcilerTestSuite. The key of the map is the name of each test
case. This has a couple key advantages:
- the name of every test case is guaranteed unique by the compiler
- editor code folding will show the name and hide the detail

One downside is that the test cases will execute in random order as
iteration over a map is random in go.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>