Infer SemanticEquals method
The SemanticEquals method is deprecated and no longer used by the
ResourceManager, ChildReconciler and AggregateReconciler. The role of
the equals method is now filled by comparing the output of the
MergeBeforeUpdate method to see if there is a meaningful change to that
resource.

The SemanticEquals method caused confusing behavior if it and the
MergeBeforeUpdate methods fell out of sync, as resources would not be
updated because the SemanticEquals method returned false, even when the
MergeBeforeUpdate method would have made a meaningful mutation.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>