Allow Halting SubReconciler processing without erroring
A new sentinel error HaltSubReconcilers can be returned from any
SubReconciler to halt the normal processing of SubReconcilers and return
to the Resource/Aggregate reconciler. This error is not returned from
the root reconciler, allowing normal operations to continue like
updating the status for a ResourceReconciler.

HaltSubReconcilers is useful when a Sequence of reconcilers cannot
continue without additional state, but the reconciler as a whole is not
in error.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>