Refine test case Prepare and Cleanup methods
This is a breaking API change, existing uses of the Prepare and Cleanup
methods were half baked and uncommon. While the API signature is
changing the existing capabilities are preserved.

The Prepare method signature now takes the form:

```
func (t *testing.T, ctx context.Context, tc *XTestCase) (context.Context, error)
```

The provided context may be augmented by the Prepare method, and then
must be returned. The XTestCase is one of `ReconcilerTestCase`,
`SubReconcilerTestCase` or `AdmissionWebhookTestCase`. The method is
called before any other state for the test has been allocated, this
means the test case may be manipulated.

The Cleanup method signature now takes the form:

```
func (t *testing.T, ctx context.Context, tc *XTestCase) error
```

Values set on the context in the prepare method are available here, as
well as any values stashed in the test case metadata.

This change also consistently set a test logr into the context. This
will show reconciler logs for test failures. This behavior was already
in place for AdmissionWebhookTestCases. The context also now has a
deadline set if the test is executed with a timeout.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>