Allow test cases to interact with client builder
Controller runtime now supports list queries using field indexes with
the fake client. If a fake client makes a request using a field index
that is not registered, it will return an error.

The fake client was previously created with a scheme and an initial set
of objects. It now can be created with a func that can interact with the
client builder which is where a field index can be registered. The
`WithClientBuilder` field is optional on each test case struct and
ExpectConfig, allowing the fake client to be further configured.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>