Drop CastResource#Type
The Type field for CastResource is completely redundant with the generic
types. We can also go a step further and avoid converting the resource
when the CastType is an interface, because we know all concrete types
are already compatible and only need to be cast, saving a lot of
overhead.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>