Extend generics into SubReconcilerTestCase Resource and ExpectedResource
This means that the value of Resource and ExpectedResource must be the
same type as the reconciler handles.

No longer can a die be passed directly and unwrapped inside the test.
While this behavior was extremely convenient, it created expectations
that may not be true in other cases and depends on DeepCopyObject on the
die returning a different type than the method was called on.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>