Indicate which resources have a status sub-resource
An upstream change in the fake client now distinguishes between
patch/update requests to resource with a status sub-resource and those
without.

Requests made to a resource with the status sub-resource at the main
resource are not able to mutate the status, and requests made to the
status sub-resource are only able to modify the status.

Requests using the status sub-resource to types that have not indicated
they support the status sub-resource will return a not found error.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>