Allow ChildReconciler to stash in ReflectChildStatusOnParent
ChildReconciler#ReflectChildStatusOnParent now accepts a context object
so the values from the child resource can be stashed for subsequent
reconcilers. This is common when the child needs to be referenced by
another resource, but it is also not desired to expose the child
directly on the parent resource.

This is a breaking change, although it is easy to adapt to.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>