Respect child desired controlling owner reference
When a desired child resource has a controlling owner reference defined,
we now respect that value and no longer attempt to redefine the
reference.

This can be useful when one of the default values needs to have a
different value, or the parent resource type is not defined in the
scheme.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>