Allow errors to be quiet
ErrQuiet errors are not logged within reconciler-runtime and do not
trigger events to be recorded. This behavior is conventional rather than
technical. An implementor may choose to ignore these conventions when it
fits their purpose. Quite is not silent.

Also redefines ErrHaltSubReconcilers to wrap ErrQuiet.
HaltSubReconcilers is deprecated in favor of ErrHaltSubReconcilers.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>