Switch ResourceManager update cache to JSON Patch
It was previously using JSON Merge Patch, which will recreate structures
that are trying to be deleted. JSON Patch will fail to apply the patch
if the structure being patched doesn't exist.

For example, if a reconciler defined a volume on a Deployment on
create/update a mutating webhook will default the file permissions. If
the reconciler later needs to remove that volume, the defaulted file
permissions will no longer be re-applied.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>