Allow resources to contain unexported fields
Using unexported fields in Kubernetes resources is not common, but does
happen. Previously, these fields would cause cmp.Diff to panic. Now, we
ignore the content of unexported fields when computing a diff. As the
diff is for logging and test assertions this is relatively safe.

Semantic equality is used for detecting when a managed resource needs to
be updated. Custom equality func can be defined separately as needed.