List children exactly once for ChildSetReconciler
Previously, the ChildSetReconciler would list children once for the
set and again for each child being reconciled. At best this was
inefficient, at worst it introduced subtile issues when the content of
the listing changed over the course of the reconcile.

Now the content of the list from the ChildSetReconciler is reused by
each child that is reconciled.

Signed-off-by: Scott Andrews <scott@andrews.me>